Added: market orders now have a temp entry limit price of the latest bar Fixed: issue with insight __str__ func Todo: Fix alpaca close position function -> Not Found issue

+++ OlympusTrader/broker/base_broker.py
@@ -49,7 +49,7 @@ class BaseBroker(abc.ABC):
     
     @override
     @abc.abstractmethod
-    def close_position(self, symbol: str, qty: int = None, percent: int = None):
+    def close_position(self, symbol: str, qty: int = None, percent: float = None):
         pass
 
     @override
